bugfix: fix type safety bugs in RAG MCP server (topk type assertion, Scores bounds check, nil vector guard)
Fixes three bugs in the RAG MCP server:

1. tools.go HandleSearch: topk.(int) always fails because JSON decodes
   numbers as float64. Changed to topk.(float64) then convert to int.

2. vectordb/milvus.go SearchDocs: result.Scores[i] accessed without
   bounds check when i >= len(result.Scores). Added bounds check to
   break out of loop if Scores is shorter than ResultCount. This is
   HIGH severity since golang-filter has no recover and panics crash
   the Envoy worker.

3. vectordb/milvus.go AddDoc: len(vectors[0]) used without nil guard.
   If first document has nil Vector, dimension is 0 which is invalid for
   Milvus. Added check to return error if dimension is 0.

Fixes #4510

+++ plugins/golang-filter/mcp-server/servers/rag/tools.go
@@ -143,9 +143,9 @@ func HandleSearch(ragClient *RAGClient) common.ToolHandlerFunc {
 		if !ok {
 			return nil, fmt.Errorf("invalid query argument")
 		}
-		topK, ok := arguments["topk"].(int)
+		topK, ok := arguments["topk"].(float64)
 		if !ok {
-			topK = ragClient.config.RAG.TopK
+			topK = float64(ragClient.config.RAG.TopK)
 		}
 
 		threshold, ok := arguments["threshold"].(float64)
